refactor: service package application 으로 변경
