+++ src/main/java/roomescape/application/param/CreateReservationTimeParam.java
@@ -1,4 +1,4 @@
-package roomescape.service.param;
+package roomescape.application.param;
 
 import java.time.LocalTime;
