+++ src/main/java/roomescape/application/param/CreateThemeParam.java
@@ -1,4 +1,4 @@
-package roomescape.service.param;
+package roomescape.application.param;
 
 public record CreateThemeParam(String name, String description, String thumbnail) {
 }
