+++ src/main/java/roomescape/application/param/LoginParam.java
@@ -1,4 +1,4 @@
-package roomescape.service.param;
+package roomescape.application.param;
 
 public record LoginParam(String email, String password) {
 }
