+++ src/main/java/roomescape/application/result/AvailableReservationTimeResult.java
@@ -1,4 +1,4 @@
-package roomescape.service.result;
+package roomescape.application.result;
 
 import java.time.LocalTime;
