+++ src/main/java/roomescape/presentation/MemberController.java
@@ -5,9 +5,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.application.MemberService;
+import roomescape.application.result.MemberResult;
 import roomescape.presentation.response.MemberResponse;
-import roomescape.service.MemberService;
-import roomescape.service.result.MemberResult;
 
 @RestController
 @RequestMapping("/members")
