+++ src/main/java/roomescape/presentation/ThemeController.java
@@ -10,10 +10,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.application.ThemeService;
+import roomescape.application.result.ThemeResult;
 import roomescape.presentation.request.CreateThemeRequest;
 import roomescape.presentation.response.ThemeResponse;
-import roomescape.service.ThemeService;
-import roomescape.service.result.ThemeResult;
 
 @RestController
 @RequestMapping("/themes")
