+++ src/main/java/roomescape/presentation/request/LoginRequest.java
@@ -1,6 +1,6 @@
 package roomescape.presentation.request;
 
-import roomescape.service.param.LoginParam;
+import roomescape.application.param.LoginParam;
 
 public record LoginRequest(String email, String password) {
     public LoginParam toServiceParam() {
