+++ src/main/java/roomescape/presentation/response/AvailableReservationTimeResponse.java
@@ -1,7 +1,7 @@
 package roomescape.presentation.response;
 
 import java.time.LocalTime;
-import roomescape.service.result.AvailableReservationTimeResult;
+import roomescape.application.result.AvailableReservationTimeResult;
 
 public record AvailableReservationTimeResponse(Long timeId, LocalTime startAt, boolean booked) {
     public static AvailableReservationTimeResponse from(AvailableReservationTimeResult availableReservationTimeResult) {
